Removed if-checks from file-controller
Made a lot of tests fail

+++ src/main/java/io/fi0x/languagegenerator/service/FileService.java
@@ -3,10 +3,10 @@ package io.fi0x.languagegenerator.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.fi0x.languagegenerator.logic.converter.LanguageConverter;
 import io.fi0x.languagegenerator.logic.dto.LanguageData;
-import io.fi0x.languagegenerator.logic.dto.LanguageJson;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
@@ -17,11 +17,19 @@ import java.nio.file.Files;
 @Service
 public class FileService
 {
-    public Resource getLanguageFile(LanguageData languageData) throws IllegalArgumentException, IOException
+    public Resource getLanguageFile(LanguageData languageData) throws IllegalArgumentException, IOException, IllegalAccessException
     {
         log.trace("getLanguageFile() called for languageData={}", languageData);
+
         if (languageData == null || languageData.getId() == null)
             throw new IllegalArgumentException("LanguageData must not be null");
+
+        String currentUser = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!currentUser.equals(languageData.getUsername()))
+        {
+            log.info("User '{}' tried to get language file with languageId={}, but does not have access to it", currentUser, languageData.getId());
+            throw new IllegalAccessException("User is not allowed to download the selected language");
+        }
 
         File actualFile = Files.createTempFile(languageData.getId().toString(), ".json").toFile();
         actualFile.deleteOnExit();
@@ -30,17 +38,5 @@ public class FileService
         mapper.writeValue(actualFile, LanguageConverter.convertToJson(languageData));
 
         return new UrlResource(actualFile.toURI());
-    }
-
-    public boolean isFileValid(LanguageJson languageJson)
-    {
-        log.trace("isFileValid() called for languageJson={}", languageJson);
-        if (languageJson.getNameLengths() == null || languageJson.getSpecialCharacterLengths() == null
-                || languageJson.getSpecialCharacterChance() < 0 || languageJson.getSpecialCharacterChance() > 1)
-        {
-            return false;
-        }
-
-        return languageJson.getNameLengths().length == 2 && languageJson.getSpecialCharacterLengths().length == 4;
     }
 }
